API: Split indexing API out of c4Query.h into c4Index.h
No code changes at all.
