+++ Replicator/tests/ReplicatorSGTest.cc
@@ -21,6 +21,7 @@
 #include "c4BlobStore.h"
 #include "c4Document+Fleece.h"
 #include "c4DocEnumerator.h"
+#include "c4Index.h"
 #include "c4Query.h"
 #include "Stopwatch.hh"
 #include "StringUtil.hh"
